Integrate separate sgp code with XnatUtils,launcher,processors_v3
