+++ dax/processors_v3.py
@@ -9,6 +9,7 @@ from uuid import uuid4
 from datetime import date
 
 from . import XnatUtils, task
+from .XnatUtils import XnatUtilsError
 from . import yaml_doc
 from .errors import AutoProcessorError
 from .task import NeedInputsException
@@ -1372,3 +1373,676 @@ def parse_procversion(yaml_file):
     tmp = os.path.splitext(yaml_file)[0]
     tmp = tmp.rsplit('_v')[1]
     return tmp
+
+
+class SgpProcessor(Processor_v3):
+    """Processor class for SGP v3 YAML files"""
+
+    def __init__(
+        self,
+        xnat,
+        yaml_file,
+        user_inputs=None,
+        singularity_imagedir=None,
+        job_template='~/job_template.txt',
+    ):
+        super(SgpProcessor, self).__init__(
+            xnat,
+            yaml_file,
+            user_inputs=user_inputs,
+            singularity_imagedir=singularity_imagedir,
+            job_template=job_template)
+
+        self.xsitype = "proc:subjgenprocdata"
+
+    def get_assessor(self, xnat, subject, inputs, project_data):
+        proctype = self.get_proctype()
+        dfa = project_data['sgp']
+        dfa = dfa[(dfa.SUBJECT == subject) & (dfa.PROCTYPE == proctype)]
+        dfa = dfa[(dfa.INPUTS == inputs)]
+
+        if len(dfa) > 0:
+            # Get the info for the assessor
+            info = dfa.to_dict('records')[0]
+
+            LOGGER.debug('matches existing:{}'.format(info['ASSR']))
+
+            # Get the assessor object
+            assr = xnat.select_experiment(
+                info['PROJECT'],
+                info['SUBJECT'],
+                info['ASSR'])
+        else:
+            LOGGER.debug('no existing assessors found, creating a new one')
+            (assr, info) = self.create_assessor(
+                xnat,
+                project_data.get('name'),
+                subject,
+                inputs)
+
+            LOGGER.debug('created:{}'.format(info['ASSR']))
+
+        return (assr, info)
+
+    def _read_yaml(self, yaml_file):
+        """
+        Method to read the processor
+
+        :param yaml_file: path to yaml file defining the processor
+        """
+        LOGGER.debug('reading processor from yaml:{}'.format(yaml_file))
+        doc = yaml_doc.YamlDoc().from_file(yaml_file).contents
+
+        # NOTE: we are assuming this yaml has already been validated
+
+        # Set version from yaml
+        self.procyamlversion = doc.get('procyamlversion')
+
+        # Set requirements from Yaml
+        reqs = doc.get('requirements')
+        self.walltime_str = reqs.get('walltime', '0-2')
+        self.memreq_mb = reqs.get('memory', '16G')
+        self.ppn = reqs.get('ppn', 1)
+        self.env = reqs.get('env', None)
+
+        # Load the command text
+        self.command = doc.get('command')
+
+        # Set Inputs from Yaml
+        inputs = doc.get('inputs')
+
+        # Handle vars
+        for key, value in inputs.get('vars', {}).items():
+            # If value is a key in command
+            k_str = '{{{}}}'.format(key)
+            if k_str in self.command:
+                self.user_overrides[key] = value
+            else:
+                if isinstance(value, bool) and value is True:
+                    self.extra_user_overrides[key] = ''
+                elif value and value != 'None':
+                    self.extra_user_overrides[key] = value
+
+        # Get xnat inputs, apply edits, then parse
+        self.xnat_inputs = inputs.get('xnat')
+        # TODO: self._edit_inputs()
+        self._parse_xnat_inputs()
+
+        # Containers
+        self.containers = []
+        for c in doc.get('containers'):
+            curc = copy.deepcopy(c)
+
+            # Set container path
+            cpath = curc['path']
+
+            if not os.path.isabs(cpath) and self.singularity_imagedir:
+                # Prepend singularity imagedir
+                curc['path'] = os.path.join(self.singularity_imagedir, cpath)
+
+            if curc.get('primary', False):
+                self.container_path = curc.get('path')
+
+            # Add to our containers list
+            self.containers.append(curc)
+
+        # Check primary container
+        if not self.container_path:
+            if len(self.containers) == 1:
+                self.container_path = self.containers[0].get('path')
+            else:
+                msg = 'multiple containers requires a primary to be set'
+                LOGGER.error(msg)
+                raise AutoProcessorError(msg)
+
+        # Outputs from Yaml
+        self._parse_outputs(doc.get('outputs'))
+
+        # Override template
+        if doc.get('jobtemplate'):
+            _tmp = doc.get('jobtemplate')
+
+            # Make sure we have the full path
+            if not os.path.isabs(_tmp):
+                # If only filename, we assume it is same folder as default
+                _tmp = os.path.join(os.path.dirname(self.job_template), _tmp)
+
+            # Override it
+            self.job_template = os.path.join(_tmp)
+
+    def build_cmds(self, assr, info, project_data, jobdir, resdir):
+        assr_label = info['ASSR']
+
+        # Make every input a list, so we can iterate later
+        inputs = info['INPUTS']
+        for k in inputs.keys():
+            if not isinstance(inputs[k], list):
+                inputs[k] = [inputs[k]]
+
+        # Find values for the xnat inputs
+        var2val, input_list = self.find_inputs(assr, inputs, project_data)
+
+        # Append other stuff
+        for k, v in self.user_overrides.items():
+            var2val[k] = v
+
+        for k, v in self.extra_user_overrides.items():
+            var2val[k] = v
+
+        # Include the assessor label
+        var2val['assessor'] = assr_label
+
+        # Handle xnat attributes
+        for attr_in in self.xnat_attrs:
+            _var = attr_in['varname']
+            _attr = attr_in['attr']
+            _obj = attr_in['object']
+            _val = ''
+
+            if _obj == 'subject':
+                _val = assr.parent().attrs.get(_attr)
+            elif _obj == 'session':
+                _val = assr.parent().attrs.get(_attr)
+                _ref = attr_in['ref']
+                _refval = [a.rsplit('/', 1)[1] for a in inputs[_ref]]
+                _val = ','.join([assr.parent().experiment(r).attrs.get(_attr) for r in _refval])
+            elif _obj == 'scan':
+                _ref = attr_in['ref']
+                _refval = [a.rsplit('/', 1)[1] for a in inputs[_ref]]
+                _val = ','.join(
+                    [assr.parent().scan(r).attrs.get(_attr) for r in _refval]
+                )
+            elif _obj == 'assessor':
+                if 'ref' in attr_in:
+                    _ref = attr_in['ref']
+                    _refval = [a.rsplit('/', 1)[1] for a in inputs[_ref]]
+                    _val = ','.join([assr.parent().assessor(r).attrs.get(_attr) for r in _refval])
+                else:
+                    _val = assr.attrs.get(_attr)
+            else:
+                LOGGER.error('invalid YAML')
+                err = 'YAML File:contains invalid attribute:{}'
+                raise AutoProcessorError(err.format(_attr))
+
+            if _val == '':
+                raise NeedInputsException('Missing ' + _attr)
+            else:
+                var2val[_var] = _val
+
+        # Build the command text
+        dstdir = os.path.join(resdir, assr_label)
+        assr_dir = os.path.join(jobdir, assr_label)
+        _host = assr._intf.host
+        _user = assr._intf.user
+        cmd = self.build_text(
+            var2val, input_list, assr_dir, dstdir, _host, _user)
+
+        return [cmd]
+
+    def create_assessor(self, xnat, project, subject, inputs):
+        # returns assr pyxnat object
+        # info dict of assessor info
+
+        xnatsubject = xnat.select_subject(project, subject)
+
+        serialized_inputs = json.dumps(inputs)
+        guidchars = 8  # how many characters in the guid?
+        today = str(date.today())
+
+        # Get a unique ID
+        count = 0
+        max_count = 100
+        while count < max_count:
+            count += 1
+            guid = str(uuid4())
+            assr = xnatsubject.experiment(guid)
+
+            if not assr.exists():
+                break
+
+        if count == max_count:
+            LOGGER.error('failed to find unique ID, cannot create assessor!')
+            raise AutoProcessorError()
+
+        # Build the assessor attributes as key/value pairs
+        assr_label = '-x-'.join([
+            project,
+            subject,
+            self.proctype,
+            guid[:guidchars]])
+
+        kwargs = {
+            'label': assr_label,
+            'ID': guid,
+            'proc:subjgenprocdata/proctype': self.proctype,
+            'proc:subjgenprocdata/procversion': self.procversion,
+            'proc:subjgenprocdata/procstatus': NEED_INPUTS,
+            'proc:subjgenprocdata/date': today,
+            'proc:subjgenprocdata/inputs': serialized_inputs}
+
+        # Create the assessor
+        LOGGER.info('creating subject asssessor:{}'.format(assr_label))
+        assr.create(experiments='proc:subjgenprocdata', **kwargs)
+
+        # We keep the inputs as a dictionary in the returned info
+        # this is also how load_sgp_data behaves
+        info = {
+            'ASSR': assr_label,
+            'QCSTATUS': JOB_PENDING,
+            'XSITYPE': 'proc:subjgenprocdata',
+            'PROCTYPE': self.proctype,
+            'PROCVERSION': self.procversion,
+            'PROCSTATUS': NEED_INPUTS,
+            'INPUTS': inputs}
+
+        return (assr, info)
+
+    def parse_subject(self, subject, project_data):
+        LOGGER.debug('parsing subject:{}'.format(subject))
+        """
+        Parse a subject to determine what assessors *should* exist for
+        this processor
+        """
+
+        artefacts_by_input = self._map_inputs(
+            subject, project_data)
+        # print('artefacts_by_input=', artefacts_by_input)
+
+        parameter_matrix = self._generate_parameter_matrix(artefacts_by_input)
+        # print('parameter_matrix=', parameter_matrix)
+
+        # TODO: filter down the combinations by applying any filters
+
+        return parameter_matrix
+
+    def find_inputs(self, assr, inputs, project_data):
+        """
+        Find the files or directories on xnat for the inputs
+
+        takes an assessor, its input artefacts, its relevant sessions
+        and returns the full paths to the input files/directories
+
+        :param assr:
+        :param sessions:
+        :param assr_inputs:
+
+        :return: variable_set, input_list:
+
+        """
+        variable_set = {}
+        input_list = []
+
+        # This will raise a NeedInputs exception if any inputs aren't ready
+        verify_artefact_status(self.proc_inputs, inputs, project_data)
+
+        # print(self.variables_to_inputs.items())
+
+        # Map from parameters to input resources
+        LOGGER.debug('mapping params to artefact resources')
+        for k, v in list(self.variables_to_inputs.items()):
+            LOGGER.debug('mapping:' + k, v)
+            inp = self.proc_inputs[v['input']]
+            resource = v['resource']
+
+            # print('vinput=', v['input'])
+            # print('resource=', resource)
+            # print(inp['resources'])
+
+            # Find the resource
+            cur_res = None
+            for inp_res in inp['resources']:
+                if inp_res['varname'] == k:
+                    cur_res = inp_res
+                    break
+
+            # TODO: optimize this to get resource list only once
+            for vnum, vinput in enumerate(inputs[v['input']]):
+                # print(vnum, vinput)
+                fname = None
+                robj = get_resource(assr._intf, vinput, resource)
+
+                # Get list of all files in the resource, relative paths
+                file_list = [x._urn for x in robj.files().get('path')]
+                if len(file_list) == 0:
+                    LOGGER.debug('empty or missing resource')
+                    raise NeedInputsException('No Resource')
+
+                if 'fmatch' in cur_res:
+                    fmatch = cur_res['fmatch']
+                elif cur_res['ftype'] == 'FILE':
+                    # Default to all
+                    fmatch = '*'
+                else:
+                    fmatch = None
+
+                if 'filepath' in cur_res:
+                    fpath = cur_res['filepath']
+                    res_path = resource + '/files/' + fpath
+
+                    # Get base file name to be downloaded
+                    fname = os.path.basename(fpath)
+                elif fmatch:
+                    # Filter list based on regex matching
+                    regex = utilities.extract_exp(fmatch, full_regex=False)
+                    file_list = [x for x in file_list if regex.match(x)]
+
+                    if len(file_list) == 0:
+                        LOGGER.debug('no matching files found on resource')
+                        raise NeedInputsException('No Files')
+
+                    if len(file_list) > 1:
+                        # Multiple files found, we only support explicit
+                        # declaration of fmulti==any1, which tells dax to use
+                        # any of the multiple files. We may later support
+                        # other options
+
+                        if 'fmulti' in cur_res and cur_res['fmulti'] == 'any1':
+                            LOGGER.debug('multiple files, fmulti==any1, using first found')
+                        else:
+                            LOGGER.debug('multiple files, fmulti not set')
+                            raise NeedInputsException(artk + ': multiple files')
+
+                    # Create the full path to the file on the resource
+                    res_path = '{}/files/{}'.format(resource, file_list[0])
+
+                    # Get just the filename for later
+                    fname = os.path.basename(file_list[0])
+                else:
+                    # We want the whole resource
+                    res_path = resource + '/files'
+
+                    # Get just the resource name for later
+                    fname = resource
+
+                variable_set[k] = get_uri(assr._intf.host, vinput, res_path)
+
+                if 'fdest' not in cur_res:
+                    # Use the original file/resource name
+                    fdest = fname
+                elif len(inputs[v['input']]) > 1:
+                    fdest = str(vnum) + cur_res['fdest']
+                else:
+                    fdest = cur_res['fdest']
+
+                if 'ddest' in cur_res:
+                    ddest = cur_res['ddest']
+                else:
+                    ddest = ''
+
+                # Append to inputs to be downloaded
+                input_list.append(
+                    {
+                        'fdest': fdest,
+                        'ftype': cur_res['ftype'],
+                        'fpath': variable_set[k],
+                        'ddest': ddest,
+                    }
+                )
+
+                # Replace path with destination path after download
+                if 'varname' in cur_res:
+                    variable_set[k] = fdest
+
+        LOGGER.debug('finished mapping params to artefact resources')
+
+        return variable_set, input_list
+
+    def _parse_xnat_inputs(self):
+        # Get the xnat attributes
+        # TODO: validate these
+        self.xnat_attrs = self.xnat_inputs.get('attrs', list())
+
+        # Get the xnat edits
+        # TODO: validate these
+        self.proc_edits = self.xnat_inputs.get('edits', list())
+
+        # get sessions
+        sessions = self.xnat_inputs.get('sessions', list())
+
+        for sess in sessions:
+            sesstypes = [_.strip() for _ in sess['types'].split(',')]
+
+            # get scans
+            scans = sess.get('scans', list())
+            for s in scans:
+                name = s.get('name')
+                self.iteration_sources.add(name)
+
+                types = [_.strip() for _ in s['types'].split(',')]
+
+                resources = s.get('resources', [])
+
+                if 'nifti' in s:
+                    # Add a NIFTI resource using value as fdest
+                    resources.append(
+                        {'resource': 'NIFTI', 'fdest': s['nifti']})
+
+                if 'edat' in s:
+                    # Add an EDAT resource using value as fdest
+                    resources.append(
+                        {'resource': 'EDAT', 'fdest': s['edat']})
+
+                needs_qc = s.get('needs_qc', False)
+
+                # Consider an MR scan for an input if it's marked Unusable?
+                skip_unusable = s.get('skip_unusable', False)
+
+                # Include 'first' or 'all' matching scans as possible inputs
+                keep_multis = s.get('keep_multis', 'all')
+
+                self.proc_inputs[name] = {
+                    'sesstypes': sesstypes,
+                    'types': types,
+                    'artefact_type': 'scan',
+                    'needs_qc': needs_qc,
+                    'resources': resources,
+                    'required': True,
+                    'skip_unusable': skip_unusable,
+                    'keep_multis': keep_multis,
+                }
+
+            # get assessors
+            asrs = sess.get('assessors', list())
+            for a in asrs:
+                name = a.get('name')
+                self.iteration_sources.add(name)
+
+                types = [_.strip() for _ in a['types'].split(',')]
+                resources = a.get('resources', [])
+                artefact_required = False
+                for r in resources:
+                    r['required'] = r.get('required', True)
+                artefact_required = artefact_required or r['required']
+
+                self.proc_inputs[name] = {
+                    'sesstypes': sesstypes,
+                    'types': types,
+                    'artefact_type': 'assessor',
+                    'needs_qc': a.get('needs_qc', False),
+                    'resources': resources,
+                    'required': artefact_required,
+                }
+
+        if 'filters' in self.xnat_inputs:
+            self._parse_filters(self.xnat_inputs.get('filters'))
+
+        self._populate_proc_inputs()
+        self._parse_variables()
+
+    def _map_inputs(self, subject, project_data):
+        inputs = self.proc_inputs
+        artefacts_by_input = {k: [] for k in inputs}
+
+        # Get lists for scans/assrs for this subject
+        scans = project_data.get('scans').to_dict('records')
+        scans = [x for x in scans if x['SUBJECT'] == subject]
+        assrs = project_data.get('assessors').to_dict('records')
+        assrs = [x for x in assrs if x['SUBJECT'] == subject]
+
+        # Find list of scans/assessors that match each specified input
+        # for i, iv in list(inputs.items()):
+        for i, iv in sorted(inputs.items()):
+            if iv['artefact_type'] == 'scan':
+                # Input is a scan, so we iterate subject scans
+                # to look for matches
+                for cscan in scans:
+                    # First we try to match the session type of the scan
+                    for typeexp in iv['sesstypes']:
+                        regex = utilities.extract_exp(typeexp)
+                        if regex.match(cscan.get('SESSTYPE')):
+                            # Sess type matches, now match scan type
+                            for expression in iv['types']:
+                                regex = utilities.extract_exp(expression)
+                                if regex.match(cscan.get('SCANTYPE')):
+                                    scanid = cscan.get('ID')
+                                    if iv['skip_unusable'] and cscan.get('QUALITY') == 'unusable':
+                                        LOGGER.info(f'Excluding unusable scan {scanid}')
+                                    else:
+                                        # Get scan path, scan ID for each matching scan.
+                                        # Break if the scan matches so we don't find it again comparing
+                                        # vs a different requested type
+                                        artefacts_by_input[i].append(cscan.get('full_path'))
+                                        break
+
+            elif iv['artefact_type'] == 'assessor':
+                for typeexp in iv['sesstypes']:
+                    for cassr in assrs:
+                        regex = utilities.extract_exp(typeexp)
+                        sesstype = cassr.get('SESSTYPE')
+                        proctype = cassr.get('PROCTYPE')
+                        # print(typeexp, sesstype, proctype, iv['types'])
+                        if regex.match(sesstype) and proctype in iv['types']:
+                            #print('MATCH!')
+                            # Session type and proc type both match
+                            artefacts_by_input[i].append(cassr.get('full_path'))
+                        else:
+                            #print('nope')
+                            pass
+
+        return artefacts_by_input
+
+    def _generate_parameter_matrix(self, artefacts_by_input):
+        inputs = self.proc_inputs
+        iteration_sources = self.iteration_sources
+
+        # generate n dimensional input matrix based on iteration sources
+        all_inputs = []
+        input_dimension_map = []
+
+        # check whether all inputs are present
+        for i, iv in list(inputs.items()):
+            if len(artefacts_by_input[i]) == 0 and iv['required'] is True:
+                return []
+
+        for i in iteration_sources:
+            # find other inputs that map to this iteration source
+            mapped_inputs = [i]
+            cur_input_vector = artefacts_by_input[i][:]
+
+            # build up the set of mapped input vectors one by one based on
+            # the select mode of the mapped input
+            combined_input_vector = [cur_input_vector]
+
+            # 'trim' the input vectors to the number of entries of the
+            # shortest vector. We don't actually truncate the datasets but
+            # just use the number when transposing, below
+            min_entry_count = min((len(e) for e in combined_input_vector))
+
+            # transpose from list of input vectors to input entry lists,
+            # one per combination of inputs
+            merged_input_vector = [
+                [None for col in range(len(combined_input_vector))]
+                for row in range(min_entry_count)
+            ]
+            for row in range(min_entry_count):
+                for col in range(len(combined_input_vector)):
+                    merged_input_vector[row][col] = combined_input_vector[col][row]
+
+            all_inputs.append(mapped_inputs)
+            input_dimension_map.append(merged_input_vector)
+
+        # perform a cartesian product of the dimension map entries to get the
+        # final input combinations
+        matrix = [
+            list(itertools.chain.from_iterable(x))
+            for x in itertools.product(*input_dimension_map)
+        ]
+
+        matrix_headers = list(itertools.chain.from_iterable(all_inputs))
+
+        # rebuild the matrix to order the inputs consistently
+        final_matrix = []
+        for r in matrix:
+            row = dict()
+            for i in range(len(matrix_headers)):
+                row[matrix_headers[i]] = r[i]
+            final_matrix.append(row)
+
+        return final_matrix
+
+
+def get_scan_status(project_data, scan_path):
+    path_parts = scan_path.split('/')
+    sess_label = path_parts[6]
+    scan_label = path_parts[8]
+
+    scans = project_data.get('scans').to_dict('records')
+    for s in scans:
+        if s['SESSION'] == sess_label and s['SCANID'] == scan_label:
+            return s['QUALITY']
+
+    raise XnatUtilsError('Invalid scan path:' + scan_path)
+
+
+def get_assr_status(project_data, assr_path):
+    path_parts = assr_path.split('/')
+    sess_label = path_parts[6]
+    assr_label = path_parts[8]
+
+    assrs = project_data.get('assessors').to_dict('records')
+    for a in assrs:
+        if a['SESSION'] == sess_label and a['ASSR'] == assr_label:
+            return a['PROCSTATUS'], a['QCSTATUS']
+
+    raise XnatUtilsError('Invalid assessor path:' + assr_path)
+
+
+def verify_artefact_status(proc_inputs, assr_inputs, project_data):
+
+    # Check artefact status
+    LOGGER.debug('checking status of each artefact')
+    for artk, artv in list(assr_inputs.items()):
+        LOGGER.debug('checking status:' + artk)
+        inp = proc_inputs[artk]
+        art_type = inp['artefact_type']
+
+        if art_type == 'scan' and not inp['needs_qc']:
+            # Not checking qc status
+            continue
+
+        if art_type == 'scan':
+            # Check status of each input scan
+            for vinput in artv:
+                qstatus = get_scan_status(project_data, vinput)
+                if qstatus.lower() == 'unusable':
+                    raise NeedInputsException(artk + ': Not Usable')
+        else:
+            # Check status of each input assr
+            for vinput in artv:
+                pstatus, qstatus = get_assr_status(project_data, vinput)
+                if pstatus in OPEN_STATUS_LIST + [NEED_INPUTS]:
+                    raise NeedInputsException(artk + ': Not Ready')
+
+                if qstatus in [JOB_PENDING, REPROC, RERUN]:
+                    raise NeedInputsException(artk + ': Not Ready')
+
+                if not inp['needs_qc']:
+                    # Not checking qc status
+                    continue
+
+                if qstatus in [FAILED_NEEDS_REPROC, NEEDS_QA]:
+                    raise NeedInputsException(artk + ': Needs QC')
+
+                for badstatus in BAD_QA_STATUS:
+                    if badstatus.lower() in qstatus.split(' ')[0].lower():
+                        raise NeedInputsException(artk + ': Bad QC')
